fetch images immediately
when the fetch code was in the view, it didn't fetch until the view was
on screen, meaning you would initially see the loading image for every
view in the table. This way they fetch as soon as we have the imageUrl.

However, the *top* image is still not being fetched in time, and because
the UI is not reactive (*snort*) the uiimage doesn't refresh. You can
see this be scrolling off and back to the spot. This causes the UIImage
do be collected and redrawn - now that the image has been fetched you'll
see the spot.

I think maybe the solution is to refactor the spotview to compose a UIView.