fixes: new spot not appearing when it's your first.
This was introduced with the APN feature, and highlighted that handing
off data via segue's is kind of a pain when you want to have multiple
code paths. The alternative would have been copy/pasting code to
daisy-chaining the new spot to the first-time-spot controller *then* to
the home controller, but really, I think the intent is something more
like a "central spots store" that all controllers have access to.

I just narrowly avoided using core data ;)